+++ docs/changelog.rst
@@ -23,6 +23,7 @@ Fixed
 -----
 - Wrong dtype for columns when reading a mapping table in **setup_laimaps_from_lulc_mapping** . PR #290
 - Read/write staticgeoms if dir_input folder is present in the wflow toml file. PR #286
+- Creating subfolders for the config file of wflow in **write_config**. PR #286
 
 v0.6.0 (7 June 2024)
 ====================